fix: split auth option and add comment
